refactor: evm helpers
