+++ packages/web3-shared/evm/src/helpers/isValidChainId.ts
@@ -1,0 +1,6 @@
+import { getEnumAsArray } from '@masknet/kit'
+import { ChainId } from '../types/index.js'
+
+export function isValidChainId(chainId?: ChainId): chainId is ChainId {
+    return getEnumAsArray(ChainId).some((x) => x.value === chainId)
+}
