+++ packages/web3-shared/evm/src/helpers/parseChainId.ts
@@ -1,0 +1,8 @@
+import { isValidChainId } from './isValidChainId.js'
+import type { ChainId } from '../types/index.js'
+
+export function parseChainId(chainId: number | string): ChainId | undefined {
+    if (typeof chainId === 'string') return parseChainId(Number.parseInt(chainId, 16))
+    if (typeof chainId === 'number' && isValidChainId(chainId)) return chainId as ChainId
+    return
+}
