+++ packages/web3-shared/evm/src/index.ts
@@ -2,6 +2,7 @@ export * from './types/index.js'
 export * from './constants/index.js'
 export * from './libs/index.js'
 export * from './helpers/address.js'
+export * from './helpers/defaults.js'
 export * from './helpers/token.js'
 export * from './helpers/resolver.js'
 export * from './helpers/formatter.js'
@@ -11,6 +12,7 @@ export * from './helpers/createContract.js'
 export * from './helpers/decodeOutputString.js'
 export * from './helpers/parseStringOrBytes32.js'
 export * from './helpers/isValidDomain.js'
+export * from './helpers/isValidChainId.js'
 export * from './helpers/decodeEvents.js'
 export * from './helpers/getTransactionStatusType.js'
 export * from './helpers/getTransactionSignature.js'
